+++ login-workflow/src/screens/AccountDetailsScreen/AccountDetailsScreen.tsx
@@ -8,7 +8,7 @@ import { useTranslation } from 'react-i18next';
  * Component renders a screen with account details information for support with the application.
  * Contact information is pulled from the context passed into the workflow.
  *
- * @param {AccountDetailsScreenProps} props - props of accountDetailsScreen
+ * @param {AccountDetailsScreenProps} props - props of AccountDetailsScreen
  *
  * @category Component
  */
