+++ login-workflow/src/screens/CreatePasswordScreen/CreatePasswordScreen.tsx
@@ -9,7 +9,7 @@ import { useTranslation } from 'react-i18next';
 /**
  * The component renders a screen with the password and confirm password field for creating a new password.
  *
- * @param {CreatePasswordScreenProps} props - props of createPasswordScreen
+ * @param {CreatePasswordScreenProps} props - props of CreatePasswordScreen
  *
  * @category Component
  */
