+++ login-workflow/src/screens/EulaScreen/EulaScreenBase.tsx
@@ -19,7 +19,7 @@ import { useTranslation } from 'react-i18next';
 /**
  * Component that renders a screen displaying the EULA and requests acceptance via a checkbox.
  *
- * @param {EulaScreenProps} props - props of eulaScreen
+ * @param {EulaScreenProps} props - props of EUlA screen base component
  *
  * @category Component
  */
