+++ login-workflow/src/screens/VerifyCodeScreen/VerifyCodeScreenBase.tsx
@@ -14,7 +14,7 @@ import ErrorManager from '../../components/Error/ErrorManager';
  * Component that renders a screen that prompts a user to enter the confirmation code
  * that was sent to the email address that they used to register.
  *
- * @param {VerifyCodeScreenProps} props - props of verifyCodeScreen base component
+ * @param {VerifyCodeScreenProps} props - props of VerifyCodeScreen base component
  *
  * @category Component
  */
